refactored client stream call to proto point to  ReadableStream
