+++ lib/calls/generateDuplex.js
