+++ lib/compose.js
@@ -1,7 +1,7 @@
 const generateUnaryCall = require("./calls/unaryCall");
-const generateClientStreamCall = require("./calls/clientStreamCall");
+const generateClientStreamCall = require("./calls/generateClientStreamCall");
 const ServerStreamCall = require("./calls/serverStreamCall");
-const generateDuplex = require("./calls/duplexCall");
+const generateDuplex = require("./calls/generateDuplex");
 const getMethodType = require("./utils/getMethodType");
 
 module.exports = function compose(
